Big refactor!
