+++ _doc/api/saul.waveform.helpers.rst
@@ -1,0 +1,6 @@
+``saul.waveform.helpers``
+=========================
+
+.. automodule:: saul.waveform.helpers
+   :members:
+   :undoc-members:
