+++ _doc/api/saul.waveform.rst
@@ -6,4 +6,5 @@
 .. toctree::
    :maxdepth: 1
 
+   saul.waveform.helpers
    saul.waveform.stream.Stream
