+++ saul/waveform/__init__.py
@@ -2,4 +2,5 @@
 Contains tools for working with waveform data.
 """
 
+from saul.waveform.helpers import get_availability
 from saul.waveform.stream import Stream
